% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix_data.R
\name{row_names}
\alias{row_names}
\alias{row_names<-}
\alias{col_names}
\alias{col_names<-}
\alias{nrows}
\alias{ncols}
\alias{is_matrix_data}
\alias{is_md_tbl}
\title{Helper function}
\usage{
row_names(md)

row_names(md) <- value

col_names(md)

col_names(md) <- value

nrows(md)

ncols(md)

is_matrix_data(md)

is_md_tbl(md)
}
\arguments{
\item{md}{a matrix_data object.}

\item{value}{a valid value for dimension names.}
}
\description{
Helper function for matrix_data object
}
\author{
Hou Yun
}
