% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_data.R
\name{matrix_data}
\alias{matrix_data}
\title{Matrix data}
\usage{
matrix_data(x, row_names = NULL, col_names = NULL)
}
\arguments{
\item{x}{a list of matrix-like data.}

\item{row_names, col_names}{the name of rows and columns.}
}
\value{
a object of matrix_data
}
\description{
Matrix data
}
\author{
Hou Yun
}
